refactor!: rename main package
