+++ src/main/java/ru/ArtemSmirnov/java1/homework8/MainClass.java
@@ -1,0 +1,7 @@
+package ru.ArtemSmirnov.java1.homework8;
+
+public class MainClass {
+    public static void main(String[] args) {
+        MyWindow myWindow = new MyWindow();
+    }
+}
