+++ src/main/java/ru/ArtemSmirnov/java1/homework8/MyWindow.java
@@ -1,0 +1,21 @@
+package ru.ArtemSmirnov.java1.homework8;
+
+import javax.swing.*;
+import java.awt.*;
+
+public class MyWindow extends JFrame {
+    public MyWindow() {
+        setBounds(500,500,400,300);
+        setTitle("GridLayoutDemo");
+        setDefaultCloseOperation(WindowConstants.EXIT_ON_CLOSE);
+        JButton[] jbs = new JButton[9];
+        setLayout(new GridLayout(3, 3));
+        for (int i = 0; i < jbs.length; i++) {
+            jbs[i] = new JButton("#" + i);
+            add(jbs[i]);
+        }
+        setVisible(true);
+
+
+    }
+}
